Use of Sovereign AI processors
