+++ .gitignore
@@ -1,3 +1,2 @@
 env/
 .env
-benthos-embeddings
