+++ requirements.txt
@@ -1,0 +1,55 @@
+aiohappyeyeballs==2.4.0
+aiohttp==3.10.5
+aiosignal==1.3.1
+annotated-types==0.7.0
+anyio==4.4.0
+attrs==24.2.0
+beautifulsoup4==4.12.3
+bs4==0.0.2
+certifi==2024.8.30
+charset-normalizer==3.3.2
+confluent-kafka==2.5.3
+dataclasses-json==0.6.7
+distro==1.9.0
+dnspython==2.6.1
+frozenlist==1.4.1
+greenlet==3.0.3
+h11==0.14.0
+httpcore==1.0.5
+httpx==0.27.2
+idna==3.8
+jiter==0.5.0
+jsonpatch==1.33
+jsonpointer==3.0.0
+langchain==0.2.16
+langchain-community==0.2.16
+langchain-core==0.2.38
+langchain-mongodb==0.1.8
+langchain-openai==0.1.23
+langchain-text-splitters==0.2.4
+langsmith==0.1.114
+marshmallow==3.22.0
+multidict==6.0.5
+mypy-extensions==1.0.0
+nest-asyncio==1.6.0
+numpy==1.26.4
+openai==1.43.0
+orjson==3.10.7
+packaging==24.1
+pydantic==2.8.2
+pydantic_core==2.20.1
+pymongo==4.8.0
+python-dotenv==1.0.1
+PyYAML==6.0.2
+regex==2024.7.24
+requests==2.32.3
+sniffio==1.3.1
+soupsieve==2.6
+SQLAlchemy==2.0.34
+tenacity==8.5.0
+tiktoken==0.7.0
+tqdm==4.66.5
+typing-inspect==0.9.0
+typing_extensions==4.12.2
+urllib3==2.2.2
+yarl==1.9.11
